+++ src/view/ContractView/ContractsTable/index.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { ContractsTable } from '@/view/ContractView/ContractsTable/ContractsTable'
 import { UserContractDetails } from '@/domain'
 import { ContractTableItem } from '@/domain/wizard/ContractTableItem'
-import { FiltersInput, FiltersInputProps } from '../SearchInput'
+import { FiltersInput, FiltersInputProps } from '../FiltersInput'
 import { Box, Button, Typography } from '@mui/material'
 import Link from 'next/link'
 import { ROUTES } from '@/constants'
